Structure Hierarchy and Structured Logger Modification Function
